refactor: 重构UserAgent转换方式
