+++ spring-extension-context/src/main/java/com/livk/context/useragent/UserAgentConverter.java
@@ -1,12 +1,9 @@
 /*
  * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  *       https://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,22 +11,16 @@
  * limitations under the License.
  */
 
-package com.livk.context.useragent.support;
+package com.livk.context.useragent;
 
 import com.livk.context.useragent.domain.UserAgent;
-import org.springframework.core.convert.converter.Converter;
 import org.springframework.http.HttpHeaders;
-import org.springframework.lang.NonNull;
 
 /**
  * @author livk
  */
-public interface UserAgentConverter extends Converter<HttpHeaders, UserAgent> {
+public interface UserAgentConverter {
 
-	@NonNull
-	@Override
-	default <U> Converter<HttpHeaders, U> andThen(@NonNull Converter<? super UserAgent, ? extends U> after) {
-		throw new UnsupportedOperationException();
-	}
+	UserAgent convert(HttpHeaders headers);
 
 }
